backend/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-filter==24.3
djangorestframework==3.15.2
Markdown==3.7
mysqlclient==2.2.7
sqlparse==0.5.3
django-simple-history==3.8.0
python-dotenv==1.0.1
django-extensions
django-cors-headers
djangorestframework-simplejwt
django-cors-middleware
pandas
